+++ src/app/admin/categories/components/data-table.tsx
@@ -21,10 +21,18 @@ import { useEffect, useState } from 'react'
 import { CategoriesDataTableColumns } from './columns';
 import { CourseCategoryData } from '@/utils/types'
 import { listAllCourseCategories } from '@/actions/category/list'
+import { deleteCategory } from '@/actions/category/register'
+import Modal from '@/components/modal'
+import { CheckIcon, XIcon } from '@/components/svg/certifit-icons'
 
 export function CategoryListDataTable() {
     const [data, setData] = useState<CourseCategoryData[]>([]);
     const [isLoading, setIsLoading] = useState(false);
+    const [isOpen, setIsOpen] = useState(false);
+    const [actionResult, setActionnResult] = useState({
+        success: false,
+        message: ''
+    });
 
     // useEffect para cargar datos cada vez que pageIndex o pageSize cambien
     useEffect(() => {
@@ -47,9 +55,23 @@ export function CategoryListDataTable() {
     }, []);
 
 
+    const handleDeleteCategory = async (id: number) => {
+        const result = await deleteCategory({ data: { id: id } });
+        setActionnResult({
+            success: result.success ?? false,
+            message: result.message ? result.message : ''
+        });
+        setIsOpen(true);
+    };
+
+    const handleModalClose = () => {
+        setIsOpen(false);
+        window.location.reload();
+    };
+
     const table = useReactTable({
         data,
-        columns: CategoriesDataTableColumns(),
+        columns: CategoriesDataTableColumns({ onDelete: handleDeleteCategory }),
         // Indicamos que la paginación es manual
         manualPagination: false,
         // Pasamos el estado de paginación actual
@@ -142,7 +164,30 @@ export function CategoryListDataTable() {
                     {'>>'}
                 </Button>
             </div>
+            {/* Modal de resultado */}
+            <div>
+                <Modal open={isOpen} setOpen={handleModalClose} closeButton={false}>
+                    <div>
+                        <div className={cn('sm:flex sm:items-start')}>
+                            <div className={cn('text-center sm:text-left w-full')}>
+                                <div className={cn('grid px-4 py-3 justify-items-center items-center sm:px-6')}>
+                                    <div className={cn('rounded-2xl flex items-center justify-center w-[89px] h-[89px] bg-[#EBF9EE]')}>
+                                        <div className={cn('rounded-2xl flex items-center justify-center w-[65px] h-[65px] bg-[#CEF4D7]')}>
+                                            {actionResult.success ? <CheckIcon width={56} height={57} className={cn('text-5xl text')} /> : <XIcon width={56} height={57} className={cn('text-5xl text')} />}
+                                        </div>
+                                    </div>
+                                </div>
+                                <div className={cn('grid px-4 py-3 justify-items-center items-center sm:px-6')}>
+                                    <p>Resultado</p>
+                                </div>
+                                <div className={cn('grid px-4 bg-[#F5F8FE] py-3 justify-items-center items-center sm:px-6')}>
+                                    <p>{actionResult.message}</p>
+                                </div>
+                            </div>
+                        </div>
+                    </div>
+                </Modal>
+            </div>
         </div>
-
     )
 }